+++ tools/xdhq/wrappers/NJS/XDHq.js
@@ -165,7 +165,7 @@ class XDH {
 	}
 	setTimeout( delay, action, callback )
 	{
-		call(this, "SetTimeout_1", types.VOID, 2, delay, action, 0, callback);
+		call(this, "SetTimeout_1", types.VOID, 2, delay.toString(), action, 0, callback);
 	}
 	dressWidgets(id, callback) {
 		call(this, "DressWidgets_1", types.VOID, 1, id, 0, callback);